refactor: reorganize timestamp
