+++ crates/proof-of-sql/src/base/time/mod.rs
@@ -1,2 +1,4 @@
-/// Stores all functionality relelvant to timestamps
+/// Stores all functionality relelvant to timestamps and time units
 pub mod timestamp;
+/// Stores functionality relevant to timezones
+pub mod timezone;
